script-python/requirements.txt
requests==2.9.1
kafka-python==1.4.4